requirements.txt
bs4
mistletoe
gitpython
tqdm
colorama
